move shared classes to new package infrastructure
