+++ src/main/java/Infrastructure/Connection.java
@@ -1,4 +1,4 @@
-package echoserver;
+package Infrastructure;
 
 import java.io.IOException;
