+++ src/main/java/Infrastructure/SocketConnection.java
@@ -1,11 +1,11 @@
-package echoserver;
+package Infrastructure;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.Socket;
 
-public class SocketConnection implements Connection{
+public class SocketConnection implements Connection {
     private final Socket socket;
 
     public SocketConnection(Socket socket) {
