+++ src/main/java/Main.java
@@ -1,22 +1,13 @@
-import echoserver.EchoServer;
-import echoserver.Echoer;
-import echoserver.Listener;
-import echoserver.Logger;
+import Infrastructure.Connection;
+import Infrastructure.Listener;
 
 import java.net.ServerSocket;
-import java.util.concurrent.Executor;
-import java.util.concurrent.Executors;
 
 public class Main {
     public static void main(String[] args) throws Exception {
         ServerSocket serverSocket = new ServerSocket(5000);
-        Executor executor = Executors.newCachedThreadPool();
-        Logger logger = new Logger(System.out);
-        String welcomeMessage = "Welcome.  Enter a text to echo.  Press control + C to end.";
-        Echoer echoer = new Echoer(welcomeMessage);
-
-        try(Listener listener = new Listener(serverSocket);) {
-            new EchoServer(listener, executor, echoer, logger).start();
-        }
+        Listener listener = new Listener(serverSocket);
+        Connection connection = listener.listen();
+        connection.close();
     }
 }
