+++ src/main/java/echoserver/Echoable.java
@@ -1,5 +1,7 @@
 package echoserver;
 
+import Infrastructure.Connection;
+
 import java.io.IOException;
 
 public interface Echoable {
