+++ src/test/java/echoserver/EchoServerTest.java
@@ -1,5 +1,7 @@
 package echoserver;
 
+import Infrastructure.Connection;
+import Infrastructure.Listenable;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
